[elsum] missing colon, fix but didn't fix problem.
- really though I had found the problem. Makes sense that if I found all
these missing colons. You would think aha found the problem right? Well
you would be wrong. All thou this is a serious problem it did nothing
for the immediate problem of the file starting up showing everything
nothing folded or hidden.
- Would makes sense then when one finds these colons missing that you
have solved the mystery. But after fixing this it did nothing that I
could tell. So when this happen and why it did nothing is beyond me.