[elfeed] sort order to ascending.
- this sorting order makes more sense to me since it is easier to move
down in elfeed and more difficult to move up. It like reading a book
backwards. I get the updated version of a story that is further down the
list but can't filter out the stories that have been updated. So the
whole thing just seems to work better in the order. My opinion.