[org-capture-templates] made into literary file.
- Well it bothered me enough I made it an org file that is made up of
separate org-src-blocks for each org-template.
- I am hoping it makes editing them easier. Not sure how it will work?
- the one worry is keeping indent proper and balanced parentheses.
- There is only one way to find out if this is useful or not and to me
that was just doing it and trying it out. If not no big deal. I would
still walk away with more knowledge then I had before and know what not
to do as well. So worth doing I thinking even if to prove why not to do it.