[elfeed-summary] scrap tangling file, reformat.
- Using a literate programming with this was a not useful at all. It did
not work for editing, there was no good way to keep the formatting. So I
will chalk it up to a learning experience. Now I know first hand how to
do it, I just should not have picked a variable to split up. Well now I
know, it sounded good on paper?