[evilem] seemed buggy,changed commands,bad binding
- Have no idea when this bad keybinding showed up or why I made it?
makes no sense.
- This started off as easy motion next line would cause a error?
did not understand why. Decided to change commands to a avy command that
did the same thing.
- That was when I discovered the bad key map. For some reason which I
have no idea why I put C-e to nil and don't know why I did this. Should
look in the logs but I will just get rid of it for now.
- Don't really see how this could have effected 'easy-motion' so will
leave the avy commands in for now.