[fix] dired + ready-player delete file error.
- Come across this issue quit a while back. The issue was if I tried to
delete a file in dired. I would get an error telling me I needed to
toggle off ready-player.
- This worked but was not ideal. I figured there should be a simple way
to work around this.
- after multiple attempts not getting what I wanted. I finally searched
the issues page on git and found the answer. 'delete-by-moving-to-trash'
was the problem.
- So by removing this, which is not a problem since I did not use it
much at all. The problem was fixed and I could put things back.
- While doing all of this. I noticed the load time difference when it
was not being loaded.
- This promted me to try loading different and added the :after-call
- This allowed it to load after a find-file function. Saving start up
load time. So I will call this fixed.