[evilem] removed.
- many reasons why this should be removed, the commands were not that
useful, most had easier bindings already.
- was needed. This forced me to try and make sense of all the different
ways one can use to navigate a file. Once a better picture was built I
could then see how this was redundant and was better to re-examine the
commands that were there. Found a better way that did not need a
transient. So no use keeping this.