[add package] greader.
- Lot could be said about this, like one of the reasons to do this had
to do with figurine out how to set something that was set with a
'defcustom' and 'defvar'.
- A lot of trial and error went on to figure it out. Had to rewrite the
keybindings to get it to work. Doom does not like packages that use the
"SPC" key.
- While monkeying around I decided it would be useful to have the mode
start were I would use it most. Which is elfeed-show-mode. This helped a
lot of keypresses. So added hook to start everytime elfeed-show-mode is
opened. This worked as intended.
- will keep seems useful enough. Would like a natural voice but will
work on that some other time. This works well enough.